+++ backend/src/main/java/com/team6/issue_tracker/application/comment/CommentContents.java
@@ -1,4 +1,4 @@
-package com.team6.issue_tracker.comment;
+package com.team6.issue_tracker.application.comment;
 
 import lombok.Getter;
 import lombok.Value;
