+++ backend/src/main/java/com/team6/issue_tracker/application/member/MemberController.java
@@ -1,7 +1,6 @@
-package com.team6.issue_tracker.member;
+package com.team6.issue_tracker.application.member;
 
-import com.team6.issue_tracker.member.MemberRepository;
-import com.team6.issue_tracker.member.domain.Member;
+import com.team6.issue_tracker.application.member.domain.Member;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
