+++ backend/src/main/java/com/team6/issue_tracker/application/member/MemberRepository.java
@@ -1,6 +1,6 @@
-package com.team6.issue_tracker.member;
+package com.team6.issue_tracker.application.member;
 
-import com.team6.issue_tracker.member.domain.Member;
+import com.team6.issue_tracker.application.member.domain.Member;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
