+++ backend/src/main/java/com/team6/issue_tracker/application/member/domain/Member.java
@@ -1,10 +1,9 @@
-package com.team6.issue_tracker.member.domain;
+package com.team6.issue_tracker.application.member.domain;
 
 import lombok.*;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.annotation.PersistenceCreator;
 import org.springframework.data.relational.core.mapping.Column;
-import org.springframework.data.relational.core.mapping.Embedded;
 import org.springframework.data.relational.core.mapping.Table;
 
 import javax.validation.constraints.NotNull;
