+++ backend/src/main/java/com/team6/issue_tracker/config/AwsConfig.java
@@ -1,4 +1,4 @@
-package com.team6.issue_tracker.adapter.configuration;
+package com.team6.issue_tracker.config;
 
 import com.amazonaws.auth.AWSStaticCredentialsProvider;
 import com.amazonaws.auth.BasicAWSCredentials;
