+++ backend/src/test/java/com/team6/issue_tracker/comment/CommentRepositoryTest.java
@@ -1,5 +1,7 @@
 package com.team6.issue_tracker.comment;
 
+import com.team6.issue_tracker.application.comment.Comment;
+import com.team6.issue_tracker.application.comment.CommentRepository;
 import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
